DPH-44 - debugging deployment to test environment

+++ dph.cloud/shared/ci/developer_tools/scripts/scripts/load-env-vars.sh
@@ -17,32 +17,36 @@ AWS_SSM_PARAMETER_PATHS=$2 # e.g. "path1;path2;..."
 WORKING_DIR=$3
 ENV_FILE="$WORKING_DIR/.env"
 
-rm -rf $ENV_FILE
-touch $ENV_FILE
+if [[ "$AWS_SSM_PARAMETER_PATHS" == "-1" ]]; then
+  echo "AWS_SSM_PARAMETER_PATHS variable not set"
+else
+  rm -rf $ENV_FILE
+  touch $ENV_FILE
 
-IFS=';'
+  IFS=';'
 
-for i in $AWS_SSM_PARAMETER_PATHS; do
-  echo "Loading environment variables at path: $i"
+  for i in $AWS_SSM_PARAMETER_PATHS; do
+    echo "Loading environment variables at path: $i"
 
-  AWS_SSM_PARAMS_RESULT=$(echo $(aws ssm get-parameters-by-path --path "$i" --region "$AWS_REGION" --recursive --with-decryption --output "json") | jq '.Parameters')
-  AWS_SSM_PARAMS_COUNT=$(echo $AWS_SSM_PARAMS_RESULT | jq '. | length')
+    AWS_SSM_PARAMS_RESULT=$(echo $(aws ssm get-parameters-by-path --path "$i" --region "$AWS_REGION" --recursive --with-decryption --output "json") | jq '.Parameters')
+    AWS_SSM_PARAMS_COUNT=$(echo $AWS_SSM_PARAMS_RESULT | jq '. | length')
 
-  for ((j = 0; j < $((AWS_SSM_PARAMS_COUNT * 1)); j++)); do
-    echo "Current index: $j"
+    for ((j = 0; j < $((AWS_SSM_PARAMS_COUNT * 1)); j++)); do
+      echo "Current index: $j"
 
-    PARAM_NAME=$(echo $AWS_SSM_PARAMS_RESULT | jq ".[$j].Name")
-    PARAM_VALUE=$(echo $AWS_SSM_PARAMS_RESULT | jq ".[$j].Value")
+      PARAM_NAME=$(echo $AWS_SSM_PARAMS_RESULT | jq ".[$j].Name")
+      PARAM_VALUE=$(echo $AWS_SSM_PARAMS_RESULT | jq ".[$j].Value")
 
-    KEY=$(sed -e 's/^"//' -e 's/"$//' -e 's|'"$i"'/["]*||' <<<$(echo $PARAM_NAME))
-    VALUE=$(sed -e 's/^"//' -e 's/"$//' <<<$(echo $PARAM_VALUE))
+      KEY=$(sed -e 's/^"//' -e 's/"$//' -e 's|'"$i"'/["]*||' <<<$(echo $PARAM_NAME))
+      VALUE=$(sed -e 's/^"//' -e 's/"$//' <<<$(echo $PARAM_VALUE))
 
-    PARAM=$(echo "$KEY"="$VALUE")
+      PARAM=$(echo "$KEY"="$VALUE")
 
-    echo $PARAM >>$ENV_FILE
+      echo $PARAM >>$ENV_FILE
+    done
+
+    echo "Total parameters retrieved from AWS SSM: $AWS_SSM_PARAMS_COUNT"
+
+    echo "Done loading environment variables at path: $i"
   done
-
-  echo "Total parameters retrieved from AWS SSM: $AWS_SSM_PARAMS_COUNT"
-
-  echo "Done loading environment variables at path: $i"
-done
+fi
